Moves the definition of QFunction Context
